api: Add getMirrorPoint to WorldPoint API
